WIP sparse and device array support
Intermediate save of a working state

This change generalizes the concept of a "backend" from cupy, cuda,
numpy to different array formats. The previous backends specify dense
arrays. This is now distingushed more clearly from the device class

It introduces support for sparse.pydata, scipy.sparse and
cupyx.sparse in the backend and for several UDFs. The MemoryDataSet
can now generate tiles in all formats, converting as needed, and can also simulate
datasets with specific sets of capabilities.

As a downside this change greatly increases the parameter space that has to be
tested since all array formats have subtle and not so subtle differences in their
API. That means UDFs **HAVE** to be tested for all their input formats.
Tests for other parameters such as tile shape etc may also have to be
tested where that has been relevant before.

TODO:
- Change scheduling and default cluster setup to use GPU
  workers as hybrid GPU and CPU workers so that a list of UDFs
  where some are CPU-only and some are GPU-only doesn't cause a complete CPU fallback.
  Also start more GPU workers per GPU to keep it fed, following findings by Mat
- Support for sparse input in more UDFs
- Cleanup
- Make mypy happy
- Review
- Documentation (a lot)
- Examples
- Practical tests